+++ src/game_logic/Slot.h
@@ -14,7 +14,7 @@ public:
 
     ~Slot() override = default;
 
-    bool destroy(float currentTime) override { return true; };
+    bool destroy(int currentTime) override { return true; };
 };
 
 #endif //BOMBERMAN_SLOT_H
