+++ src/game_logic/Wall.h
@@ -14,7 +14,7 @@ public:
 
     ~Wall() override = default;
 
-    bool destroy(float currentTime) override { return true; };
+    bool destroy(int currentTime) override { return true; };
 };
 
 #endif //BOMBERMAN_WALL_H
